pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-gemini"
version = "1.0.4"
description = "Discord bot to enable Google GenAI API interactions"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "John Magnus-Sharpe" },
]
dependencies = [
  "aiohttp~=3.13",
  "google-genai~=1.73",
  "Pillow~=12.1",
  "py-cord~=2.7",
  "python-dotenv~=1.2",
  "PyYAML~=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest~=9.0",
  "pytest-asyncio~=1.3",
  "ruff~=0.15",
  "pyright~=1.1",
]

[project.urls]
Repository = "https://github.com/jdmsharpe/discord-gemini"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = [
  "E501",  # line too long (formatter handles this)
]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
include = ["src"]
exclude = [".venv", "**/node_modules", "**/__pycache__", "**/.*"]
extraPaths = ["src"]
pythonVersion = "3.10"
venvPath = "."
venv = ".venv"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["discord_gemini*"]

[tool.setuptools.package-data]
"discord_gemini.config" = ["*.yaml"]
